backend/requirements.txt
Flask==2.2.5
pymongo==4.4.0
python-dotenv==1.0.0
PyJWT==2.8.0
werkzeug==2.2.3
pdfminer.six==20221105
python-docx==0.8.11
flask-cors==4.0.0
